+++ package.js
@@ -1,12 +1,12 @@
 Package.describe({
   name : 'babrahams:accounts-ldap',
   summary: 'Meteor account login via LDAP',
-  version: '0.9.10',
+  version: '0.10.2',
   git : 'https://github.com/JackAdams/meteor-accounts-ldap',
   documentation: 'README.md'
 });
 
-Npm.depends({'ldapjs' : '1.0.2', 'connect' : '2.19.3'});
+Npm.depends({'ldapjs' : '2.3.3', 'connect' : '2.19.3'});
 
 Package.onUse(function (api) {
   api.versionsFrom(['1.8.2', '2.3']);